Add personnel code to user name + Check for books of categoryId before deleting it #Refactor

+++ Server/ATA.Library.Server.Api/Controllers/api/Category/CategoryController.cs
@@ -1,10 +1,12 @@
 ﻿using ATA.Library.Server.Model.Entities.Category;
+using ATA.Library.Server.Service.Book.Contracts;
 using ATA.Library.Server.Service.Category.Contracts;
 using ATA.Library.Shared.Dto;
+using ATA.Library.Shared.Service.Exceptions;
 using AutoMapper;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Threading;
@@ -19,14 +21,16 @@ namespace ATA.Library.Server.Api.Controllers.api.Category
     public class CategoryController : BaseApiController
     {
         private readonly ICategoryService _categoryService;
+        private readonly IBookService _bookService;
         private readonly IMapper _mapper;
 
         #region Constructor Injections
 
-        public CategoryController(ICategoryService categoryService, IMapper mapper)
+        public CategoryController(ICategoryService categoryService, IMapper mapper, IBookService bookService)
         {
             _categoryService = categoryService;
             _mapper = mapper;
+            _bookService = bookService;
         }
 
         #endregion
@@ -108,7 +112,12 @@ namespace ATA.Library.Server.Api.Controllers.api.Category
             if (entity == null)
                 return NotFound($"هیچ دسته‌ای با این شناسه پیدا نشد. شناسه‌ی ارسالی = {categoryId}");
 
-            // todo: Check any book is using that categoryId
+            var hasBooks = await _bookService.GetAll()
+                .AnyAsync(b => b.CategoryId == categoryId, cancellationToken: cancellationToken);
+
+            if (hasBooks)
+                throw new DomainLogicException("ابتدا کتاب‌های موجود در این دسته را حذف نمایید");
+
 
             await _categoryService.DeleteAsync(categoryId, cancellationToken);
